Implement saving Event

+++ asp.net-mvc-event-manager/ViewModels/EventFormViewModel.cs
@@ -8,7 +8,7 @@ namespace asp.net_mvc_event_manager.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int GenreId { get; set; }
+        public byte GenreId { get; set; }
         public IEnumerable<Genre> Genres { get; set; }
     }
 }
